# turerp/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Turerp ERP API",
    "description": "Multi-tenant SaaS ERP system API\n\n## Authentication\n\nAll endpoints except `/api/v1/auth/login`, `/api/v1/auth/register`, and `/api/v1/auth/refresh` require JWT Bearer token authentication.\n\n## Rate Limiting\n\nAuthentication endpoints are rate limited to 10 requests per minute per IP address with a burst of 3 requests.\n\n## Localization\n\nAll endpoints support per-request localization via the `Accept-Language` header. Supported languages: `en` (default), `tr`.\n\n## API Versioning\n\n- `/api/v1/` - Current stable API (recommended)\n- `/api/auth/`, `/api/users/` - Legacy routes (deprecated, will be removed in v2)",
    "contact": {
      "name": "Turerp Team",
      "email": "info@turerp.com"
    },
    "license": {
      "name": ""
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/crm/campaigns": {
      "get": {
        "tags": [
          "CRM"
        ],
        "summary": "Get all campaigns",
        "operationId": "get_campaigns",
        "responses": {
          "200": {
            "description": "List of campaigns"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "post": {
        "tags": [
          "CRM"
        ],
        "summary": "Create campaign (requires admin role)",
        "operationId": "create_campaign",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCampaign"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Campaign created"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/v1/crm/campaigns/status/{status}": {
      "get": {
        "tags": [
          "CRM"
        ],
        "summary": "Get campaigns by status",
        "operationId": "get_campaigns_by_status",
        "parameters": [
          {
            "name": "status",
            "in": "path",
            "description": "Campaign status",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/CampaignStatus"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Campaigns by status"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/v1/crm/campaigns/{id}": {
      "get": {
        "tags": [
          "CRM"
        ],
        "summary": "Get campaign by ID",
        "operationId": "get_campaign",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Campaign ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Campaign found"
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/v1/crm/campaigns/{id}/status": {
      "put": {
        "tags": [
          "CRM"
        ],
        "summary": "Update campaign status (requires admin role)",
        "operationId": "update_campaign_status",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Campaign ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCampaignStatusRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Status updated"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/v1/crm/leads": {
      "get": {
        "tags": [
          "CRM"
        ],
        "summary": "Get all leads",
        "operationId": "get_leads",
        "responses": {
          "200": {
            "description": "List of leads"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "post": {
        "tags": [
          "CRM"
        ],
        "summary": "Create lead (requires admin role)",
        "operationId": "create_lead",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLead"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Lead created"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/v1/crm/leads/status/{status}": {
      "get": {
        "tags": [
          "CRM"
        ],
        "summary": "Get leads by status",
        "operationId": "get_leads_by_status",
        "parameters": [
          {
            "name": "status",
            "in": "path",
            "description": "Lead status",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/LeadStatus"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Leads by status"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/v1/crm/leads/{id}": {
      "get": {
        "tags": [
          "CRM"
        ],
        "summary": "Get lead by ID",
        "operationId": "get_lead",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Lead ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lead found"
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/v1/crm/leads/{id}/convert": {
      "post": {
        "tags": [
          "CRM"
        ],
        "summary": "Convert lead to customer (requires admin role)",
        "operationId": "convert_lead",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Lead ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConvertLeadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Lead converted"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/v1/crm/leads/{id}/status": {
      "put": {
        "tags": [
          "CRM"
        ],
        "summary": "Update lead status (requires admin role)",
        "operationId": "update_lead_status",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Lead ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLeadStatusRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Status updated"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/v1/crm/opportunities": {
      "get": {
        "tags": [
          "CRM"
        ],
        "summary": "Get all opportunities",
        "operationId": "get_opportunities",
        "responses": {
          "200": {
            "description": "List of opportunities"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "post": {
        "tags": [
          "CRM"
        ],
        "summary": "Create opportunity (requires admin role)",
        "operationId": "create_opportunity",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOpportunity"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Opportunity created"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/v1/crm/opportunities/status/{status}": {
      "get": {
        "tags": [
          "CRM"
        ],
        "summary": "Get opportunities by status",
        "operationId": "get_opportunities_by_status",
        "parameters": [
          {
            "name": "status",
            "in": "path",
            "description": "Opportunity status",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/OpportunityStatus"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Opportunities by status"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/v1/crm/opportunities/{id}": {
      "get": {
        "tags": [
          "CRM"
        ],
        "summary": "Get opportunity by ID",
        "operationId": "get_opportunity",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Opportunity ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Opportunity found"
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/api/v1/crm/pipeline-value": {
      "get": {
        "tags": [
          "CRM"
        ],
        "summary": "Get sales pipeline value",
        "operationId": "get_pipeline_value",
        "responses": {
          "200": {
            "description": "Total pipeline value"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [
    {
      "bearer_auth": []
    }
  ],
...
}